Complete the initial MVP for the homelab DNS server.
kube-rs is used to watch for all LoadBalancer services with the
k8s.r0ssd.co/dns-entry annotation. The requested domain is then pushed
into a trust-dns-server InMemoryAuthority for actually serving
authoritative DNS.

There are a ton of things we don't support (like deleting records, or
udpating records). This MVP is enough to get started though.