[controller] Reimplement main loop as a kube-rs controller
Rather than have a big monolithic main function I've split things out
and started using the controller pattern. The DNS server is spawned, and
then we tell kube-rs to start a controller loop for Service objects.

Our reconciliation loop is a bit complicated since we aren't technically
the owner of Service objects. We still use finalizers but instead of 2
cases (Apply, Cleanup) we have 3 cases.
1. A service is added/updated with our attribute. In which case the
   logic is basically the same as before.
2. A service with our attribute is deleted. Here we use kube-rs's
   finalizer support to automatically handle the cleanup.
3. A service with our attribute is updated to remove the attribute.
   This case is tricky, but we handle it mostly the same as case 2.
   Instead of automatically removing the finalizer via kube-rs we
   do that part ourselves after running the normal cleanup routine

The code is still kinda messy but generally speaking the architecture if
a bit better then before. The next step is to do some cleanups and write
the deployment manifests.