[controller] Add cname support and emit events to Service objects
The new k8s.r0ssd.co/cname-entry annotation can be used to create CNAME
records for a service. The service must have a A or AAAA record for
CNAMES to be created. This feature should make it easier to deploy
ingress controllers where you have a single LB that serves many domains.

In order to make it easier to understand what operations the home-lab
dns controller is making it now emits events. The events are published
to the Service objects with the homelab-dns annotation.