[controller] [deployment] Add compatibility flag to only use A records
While deploying the new cname support to my homelab cluster I ran into
some funky problems with ACME HTTP01 challenges. I can dig the cnames
but curl, python, go, and wget are unable to resolve them. I think it
has something to do with coredns forwarding the query directly to
homelab-dns and not consolidating the results from the Additional
section. The trust-dns-server we use is authoritative so it will not
do that consolidation.

Seemingly most dns resolvers assume they are talking to a recursive
resolver and thus do not bother to chase CNAMEs via the Additional
section. I'm not entirely sure which system is at fault. In my homelab
the DNS flow is roughly Pod -> CoreDNS -> DNSMasq -> homelab-dns.

This change fixes the problem by just not using CNAMEs at all. Which
kinda sucks, but until I can debug which resolver or server is at fault
gets my system working.